# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: UseDanfo API — Lagos public transport route crowdsourcing
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trips:
    get:
      operationId: getTrips
      tags: [trips]
      summary: List current user's trips
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
    post:
      operationId: startTrip
      tags: [trips]
      summary: Start a new GPS trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
  /trips/{id}:
    get:
      operationId: getTrip
      tags: [trips]
      summary: Get a single trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
    patch:
      operationId: updateTrip
      tags: [trips]
      summary: Update trip (append GPS points or complete it)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
  /trips/{id}/events:
    post:
      operationId: addTripEvent
      tags: [trips]
      summary: Add an event to a trip (entered bus, dropped, paid)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripEventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripEvent"
  /fare-reports:
    post:
      operationId: submitFareReport
      tags: [trips]
      summary: Submit a quick fare report (no full trip needed)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FareReportInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FareReport"
    get:
      operationId: getFareReports
      tags: [trips]
      summary: Get fare reports for a route
      parameters:
        - name: from_area
          in: query
          schema:
            type: string
        - name: to_area
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FareReport"
...
components:
  schemas:
...
    GpsPoint:
      type: object
      required: [lat, lng, ts]
      properties:
        lat:
          type: number
        lng:
          type: number
        ts:
          type: integer
          description: Unix timestamp ms
    TripInput:
      type: object
      required: [from_area, to_area]
      properties:
        from_area:
          type: string
        to_area:
          type: string
        initial_gps:
          $ref: "#/components/schemas/GpsPoint"
    TripUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [active, completed, abandoned]
        gps_points:
          type: array
          items:
            $ref: "#/components/schemas/GpsPoint"
        to_area:
          type: string
    TripEventInput:
      type: object
      required: [event_type]
      properties:
        event_type:
          type: string
          enum: [entered_bus, dropped, paid]
        area_name:
          type: ["string", "null"]
        amount:
          type: ["integer", "null"]
        gps_point:
          $ref: "#/components/schemas/GpsPoint"
    TripEvent:
      type: object
      required: [id, trip_id, event_type, created_at]
      properties:
        id:
          type: integer
        trip_id:
          type: integer
        event_type:
          type: string
        area_name:
          type: ["string", "null"]
        amount:
          type: ["integer", "null"]
        gps_point:
          type: ["object", "null"]
        created_at:
          type: string
          format: date-time
    Trip:
      type: object
      required: [id, user_id, from_area, status, created_at]
      properties:
        id:
          type: integer
        user_id:
          type: integer
        from_area:
          type: string
        to_area:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, completed, abandoned]
        gps_trace:
          type: array
          items:
            $ref: "#/components/schemas/GpsPoint"
        events:
          type: array
          items:
            $ref: "#/components/schemas/TripEvent"
        created_at:
          type: string
          format: date-time
        completed_at:
          type: ["string", "null"]
          format: date-time
    FareReportInput:
      type: object
      required: [from_area, to_area, amount]
      properties:
        from_area:
          type: string
        to_area:
          type: string
        amount:
          type: integer
        trip_id:
          type: ["integer", "null"]
    FareReport:
      type: object
      required: [id, from_area, to_area, amount, created_at]
      properties:
        id:
          type: integer
        from_area:
          type: string
        to_area:
          type: string
        amount:
          type: integer
        trip_id:
          type: ["integer", "null"]
        created_at:
          type: string
          format: date-time
...